Refactor security related code into a separate solr_security_util.go vs. sprinkled throughout the codebase.
